基本框架50% 准备写点UI

+++ headers/uid.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "commons.h"
-typedef unsigned long long int uid;
+typedef int uid;
 
 uid newuid(void);
 uid removeuid(uid id);
